Added Vampire Spawner
i added a vampire spawner.
When 45 seconds has passed one vampire spawns afterwards they spawn every 5 seconds.

+++ src/CMS/Assets/Scipts/PlayerHealth/Health.cs
@@ -14,7 +14,7 @@ public class Health : MonoBehaviour
     }
 
     [Obsolete]
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         currentHealth -= amount;
 
@@ -25,7 +25,7 @@ public class Health : MonoBehaviour
         }
     }
 
-    public void Heal(int amount)
+    public void Heal(float amount)
     {
         currentHealth += amount;
